\name{annObj}
\docType{data}
\alias{annObj}
\alias{annobj}

\title{Default Object of class annObj}

\description{
  This dataset contains an object of the class annObj.  The annObj class
  consists of anninfo objects which provide annotation information. 
}

\usage{
data(annObj)
}

\format{see \code{annotation} for more details}

\details{
  This dataset contains an object of the class annobj.  An annobj
  contains individual objects of the class anninfo. Each anninfo object
  is an annotation set. The annotation sets included in this example
  are disease genes, cancer genes, and DNA repair genes. Other
  custom annotation tracks may be included at anytime. 
  see \code{anninfo} for more information on this class.
  see \code{CancerGenes}, \code{DiseaseGenes}, and \code{DNArepairGenes}
  for details on the files used to create annotation tracks.
  
}

\keyword{datasets}
