\name{Band.Info}
\docType{data}
\alias{band.info}
\alias{Band.Info}
\alias{Band.info}

\title{Default Object of Class Band.Info}

\description{
  This dataset contains an object of the class band.info that is used as
  the default band.info object for iGGV functions.
}

\usage{
data(Band.Info)
}

\format{see \code{bandinfo} for more detail}

\details{
This dataset contains an object of the class band.info that is used as
the default band.info object for iGGV functions. It is created using the
cytoband.txt file downloaded from the UCSC Genome Bioinformatics
website.
See also \code{bandinfo}, \code{makeBandInfo}, \code{cytoBand}
}

\source{
  UCSC Genome Bioinformatics
  \url{http://hgdownload.cse.ucsc.edu/goldenPath/hg18/database/cytoBand.txt.gz}
}

\references{
  UCSC Genome Bioinformatics
  \url{http://hgdownload.cse.ucsc.edu/goldenPath/hg18/database/cytoBand.txt.gz}
}

\note{This object was created with the cytoband file from the UCSC website}

\keyword{datasets}
