\name{cytoBand}
\docType{data}
\alias{cytoBand}
\alias{cytoband}

\title{Band Information File}

\description{
  This file contains band information
}

\usage{
data(cytoBand)
}

\format{
  a tab delimited text file with columns for chromosome, start location,
  stop location, and band. 
}

\details{
  A band information file must minimally contain columns for chromosome,
  start and stop location, and band.
  see \code{bandinfo} and \code{band.info}
}

\references{
  
  Karolchik D, Kuhn, RM, Baertsch R, Barber GP, Clawson H, Diekhans M,
  Giardine B, Harte RA, Hinrichs AS, Hsu F, Miller W, Pedersen JS, Pohl
  A, Raney BJ, Rhead B, Rosenbloom KR, Smith KE, Stanke M,
  Thakkapallayil A, Trumbower H, Wang T, Zweig AS, Haussler D, Kent
  WJ. The UCSC Genome Browser Database: 2008 update. Nucleic Acids
  Res. 2008 Jan;36:D773-9
  
}

\note{see also \code{bandinfo}, \code{band.info}, \code{makeBandInfo}}

\keyword{datasets}
