\name{HB19Kv2.HG18}
\docType{data}
\alias{HB19Kv2HG18}
\alias{HB19Kv2.HG18}

\title{Human Genome 19K BAC build Mapping File}

\description{
  This file is a mapping file for the human genome 19K BAC build.
}

\usage{
data(HB19Kv2.HG18)
}

\format{
  tab delimited text file with columns for BAC name, chromosome, start
  location, stop location, central location, genomic location, band,
  mapped by, flag, and weblink to UCSC Genome Browser.
}

\details{
  mapping files must minimally contain columns for name, chromosome,
  and location.
  see also \code{mappingObj}, \code{mapping.info}, and \code{mapobj}
}

\references{
  The March 2006 human reference sequence (NCBI Build 36.1,
  Hg18). \url{http://www.ncbi.nlm.nih.gov/genome/guide/human/release_notes.html#b36}
  
  Kent WJ, Sugnet CW, Furey TS, Roskin KM, Pringle TH, Zahler AM,
  Haussler D. The human genome browser at UCSC. Genome Res. 2002
  Jun;12(6):996-1006
}

\note{see also \code{mappingObj}, \code{mapping.info}, and \code{mapobj}}

\keyword{datasets}
