\name{iGenomicViewer-package}
\docType{package}
\alias{iGenomicViewer}

\title{
  Wrapper Package to make Bioinformatics heatmaps utilizing sendplot}

\description{
  This package creates interactive bioinformatics heatmaps across a
  genomic range. It allows for the ability to create interactive
  annotation tracks and an interactive statistical or descriptive plot
  along side the heatmap.  
}

\author{Lori A. Shepherd}

\seealso{\code{\link{iGGV}}, \code{\link{iGGVtiled}}}

\references{

  Albertson, D.G., Pinkel, D. 2003 Genomic microarrays in human
  genetic disease and cancer. 12. Spec No 2: 142-152

  Chari, R., Lockwood, W.W., Lam, W.L. 2006 Computational Methods for
  the Analysis of Array Comparative Genomic Hybridization. Cancer
  Informatics 2: 48-58

  Eisen, M.B., Spellman, P.T., Brown, P.O., Botstein, D. 1998. Cluster
  analysis and display of genome-wide expression
  patterns. Proc. Natl. Acad. Sci. USA. 95:14863-14868
  
  Gaile, D.P., Shepherd, L.A., Sucheston, L., Bruno, A., Manly,
  K.F. 2009. sendplot: Took for sending interactive plots with tool-tip
  content. http://cran.r-project.org/

  Hibbs, M.A., Dirksen, N.C., Li, K., Troyanskaya,
  O.G. 2005. Visualization methods for statistical analysis of
  microarray clusters. BMC Bioinformatics 6: 115

  Yee Hwa Yang, Paquet, A., Dudoit, S. 2007. marray: Exploratory analysis
  for two-color spotted microarray data R package version 1.20.0
  http://www.maths.usyd.edu.au/u/jeany/

  Prasad, T.V., Ahson, S.I. 2006. Visualization of microarray gene
  expression data. Bioinformation 4: 141-145

  Zorn, W 2007 JavaScript, DHTML Tooltips JavaScript Cross Browser
  Library http://www.walterzorn.com/tooltip/tooltip\_e.htm
    
}

\keyword{package}

