\name{mapping.info}
\docType{data}
\alias{mapping.info}

\title{Default Object of Class mapobj}

\description{
  This dataset contains an object of the class mapobj.
}

\usage{
data(mapping.info)
}

\format{
  see \code{mapobj} for more details
}

\note{see \code{mapobj} and \code{mappingObj}}

\keyword{datasets}
